git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@4524 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ magick/image.c
@@ -4299,9 +4299,9 @@ MagickExport MagickBooleanType SyncImage(Image *image)
       index=PushColormapIndex(image,(size_t) GetIndexPixelComponent(indexes+x),
         &range_exception);
       if (image->matte == MagickFalse)
-        SetRGBPixelComponent(q,image->colormap+index)
+        SetRGBPixelComponent(q,image->colormap+(ssize_t) index)
       else
-        SetRGBOPixelComponent(q,image->colormap+index);
+        SetRGBOPixelComponent(q,image->colormap+(ssize_t) index);
       q++;
     }
     if (SyncCacheViewAuthenticPixels(image_view,exception) == MagickFalse)
