+++ magick/xwindow.c
@@ -4458,7 +4458,8 @@ static Image *XGetWindowImage(Display *display,const Window window,
               {
                 index=(IndexPacket) XGetPixel(ximage,x,y);
                 SetIndexPixelComponent(indexes+x,index);
-                SetRGBOPixelComponent(q,composite_image->colormap+index);
+                SetRGBOPixelComponent(q,composite_image->colormap+(ssize_t)
+                  index);
                 q++;
               }
               if (SyncCacheViewAuthenticPixels(composite_view,exception) == MagickFalse)
